Move PartialImportTest.java to the new testsuite
Part of: #34494
