+++ test-framework/core/src/main/java/org/keycloak/testframework/realm/RealmConfigBuilder.java
@@ -210,6 +210,11 @@ public class RealmConfigBuilder {
         return this;
     }
 
+    public RealmConfigBuilder duplicateEmailsAllowed(boolean duplicateEmailsAllowed) {
+        rep.setDuplicateEmailsAllowed(duplicateEmailsAllowed);
+        return this;
+    }
+
     /**
      * Best practice is to use other convenience methods when configuring a realm, but while the framework is under
      * active development there may not be a way to perform all updates required. In these cases this method allows
