+++ tests/base/src/test/java/org/keycloak/tests/admin/partialimport/PartialImportGroupTest.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2016 Red Hat Inc. and/or its affiliates and other contributors
+ * as indicated by the @author tags. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package org.keycloak.tests.admin.partialimport;
+
+import org.junit.jupiter.api.Test;
+import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
+
+/**
+ * Tests for the partial import endpoint in admin client.  Also tests the
+ * server side functionality of each resource along with "fail, skip, overwrite"
+ * functions.
+ *
+ * @author Stan Silvert ssilvert@redhat.com (C) 2016 Red Hat Inc.
+ */
+@KeycloakIntegrationTest
+public class PartialImportGroupTest extends AbstractPartialImportTest {
+
+    @Test
+    public void testAddGroupsFail() {
+        addGroups();
+        testFail();
+    }
+
+
+    @Test
+    public void testAddGroupsSkip() {
+        addGroups();
+        testSkip();
+    }
+
+    @Test
+    public void testAddGroupsOverwrite() {
+        addGroups();
+        testOverwrite();
+    }
+}
