+++ tests/custom-scripts/src/main/resources/scripts/default-policy.js
@@ -1,0 +1,1 @@
+$evaluation.grant();
