+++ tests/pom.xml
@@ -37,6 +37,7 @@
         <module>utils</module>
         <module>utils-shared</module>
         <module>custom-providers</module>
+        <module>custom-scripts</module>
         <module>clustering</module>
     </modules>
